requirements.txt
streamlit
Pillow
torch
torchvision
fastapi
io